+++ pyrl/creature.py
