+++ user_input/walk_mode.py
@@ -1,9 +1,9 @@
 import const.game as GAME
-import mappings as MAPPING
+import pyrl.mappings as MAPPING
 import const.directions as DIR
-from generic_algorithms import add_vector, get_vector, clockwise, anticlockwise, reverse_vector
-from generic_algorithms import clockwise_45, anticlockwise_45
-from main import io
+from pyrl.generic_algorithms import add_vector, get_vector, clockwise, anticlockwise, reverse_vector
+from pyrl.generic_algorithms import clockwise_45, anticlockwise_45
+from pyrl.main import io
 
 CORRIDOR = (False, False)
 LEFT = (True, False)
